requirement.txt
kafka-python==2.0.2
kafka-python-ng==2.2.2
praw==7.7.1
pyspark==3.5.1
websockets==12.0
Flask==3.0.3
textblob
scikit-learn
keras